Fix broken coordinate links
This war really tricky to debug. When using AstroJS in Dev mode, all links where fine. Only when we build the page, the some links where broken. When on the homepage, all links where fine; but when on a region page, the links where broken.

Turns out, I used `.reverse()` which modifies the original object. And in dev mode, that did not cause any issue for some reason. But during build, the reversed order was stored and used for the whole page. And since this special link was only used in one page, all pages after that one had the broken data.